requirements.txt
django>=5.0
djangorestframework
django-cors-headers
channels[daphne]
channels-redis
psycopg2-binary
celery[redis]
paho-mqtt
django-filter
django-celery-beat
gunicorn
dj-database-url
django-celery-results
